make custom formatters more robust in face of "weird" inputs
Some calls to core fns might throw under some circumstances.
For example javascript window object seems to be special and throws
on access to undefined property. But it does it only in the context
of custom formatters expansion. I think this has something to do with
the way how custom formatters are expanded. The code is not executed
under normal javascript context but in some special (debugger) context.