show a compile-time warning when required under non-dev builds (#37)
The solution with devtools.optional namespace didn't work in scenario
when cljs-devtools was required as a library via maven dependency.
In that case we cannot (goog/require "devtools.optional") because the \
file is not compiled/present.

There does not seem to be a way how to conditionally require something
based on compiler options. So I decided just emit a compile-time warning
to warn users and leave it as is.

People should not use cljs-devtools in :advanced mode. And ideally they
should include it via :preloads which has no effect in :advanced mode.